Moves files into namespaces + unconcernify helpers and make them normal helper modules
