+++ app/views/godmin/shared/_navigation.html.erb
@@ -13,7 +13,7 @@
 </ul>
 <ul class="nav navbar-nav navbar-right">
   <% if authentication_enabled? &&
-        admin_user.singleton_class.include?(Godmin::AdminUser) &&
+        admin_user.singleton_class.include?(Godmin::Authentication::User) &&
         admin_user_signed_in? %>
     <li class="dropdown">
       <a href="#" class="dropdown-toggle" data-toggle="dropdown">
