+++ lib/godmin/helpers/application.rb
@@ -1,0 +1,6 @@
+module Godmin
+  module Helpers
+    module Application
+    end
+  end
+end
