+++ lib/godmin/helpers/translations.rb
@@ -1,12 +1,6 @@
 module Godmin
   module Helpers
     module Translations
-      extend ActiveSupport::Concern
-
-      included do
-        helper_method :translate_scoped
-      end
-
       def translate_scoped(translate, scope: nil, default: nil, **options)
         if @resource_class
           scope ||= @resource_class.to_s.underscore
@@ -18,7 +12,7 @@ module Godmin
           default
         ]
 
-        view_context.translate(defaults.shift, default: defaults, **options)
+        t(defaults.shift, default: defaults, **options)
       end
     end
   end
